pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "universal-file-parser"
version = "1.0.0"
description = "Universal file parser for parsing, converting and analyzing various file types"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [{name = "Developer", email = "dev@example.com"}]
keywords = ["parser", "file-parser", "json", "yaml", "csv", "xml", "excel"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "openpyxl>=3.1.0",
    "odfpy>=1.4.1",
    "pyyaml>=6.0",
    "toml>=0.10.2",
    "tomli>=2.0.1",
    "click>=8.1.0",
    "chardet>=5.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-cov>=4.1.0", "flake8>=6.0.0", "mypy>=1.0.0"]

[project.urls]
Homepage = "https://github.com/yourusername/universal-file-parser"

[project.scripts]
parser = "parser.cli:main"
